Refactor plugin to use bootstrap file following WordPress standards
